+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.5.0] - 2025-10-14
+
 ### Changed
 
 - BREAKING: Make build context proxy create the context internally
@@ -33,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial public release
 
-[unreleased]: https://github.com/fastly/compute-js-context/compare/v0.4.2...HEAD
+[unreleased]: https://github.com/fastly/compute-js-context/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/fastly/compute-js-context/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/fastly/compute-js-context/compare/v0.4.1...v0.4.2
 [0.4.1]: https://github.com/fastly/compute-js-context/compare/v0.2.0...v0.4.1
 [0.2.0]: https://github.com/fastly/compute-js-context/releases/tag/v0.2.0